services/bsl-agent/requirements.txt
# BSL Agent Dependencies
# British Sign Language gloss translation service

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Configuration & Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Observability
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-httpx>=0.42b0
opentelemetry-exporter-otlp>=1.21.0
prometheus-client>=0.19.0

# Environment
python-dotenv>=1.0.0

# Testing (dev dependencies)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.2
